// backend/controllers/http/API/middlewares/permissions.ts
export const setPermission = (permission: string) => (req: any, _: PXIOHTTP.Response, next: Next) => {
  req.permission = permission
  next()
}

export enum PERMISSIONS {
  AUTH_LOGIN = 'auth_login',
  AUTH_LOGOUT = 'auth_logout',
  AUTH_STATUS = 'auth_status',
  AUTH_CHANGE = 'auth_change',
  FS_CHANGE = 'fs_change',
  FS_COPY = 'fs_copy',
  FS_LS = 'fs_ls',
  FS_MKDIR = 'fs_mkdir',
  FS_RM = 'fs_rm',
  FS_MOVE = 'fs_move',
  FS_RENAME = 'fs_rename',
  FS_UPLOAD = 'fs_upload',
  FS_DOWNLOAD = 'fs_download',
  PROFILE_INFO = 'profile_info',
  PROFILE_APP_LIST = 'profile_app_list',
  PROFILE_READ_CONFIG = 'profile_read_config',
  PROFILE_WRITE_CONFIG = 'profile_write_config',
  PROFILE_UPDATE = 'profile_update',
  PROFILE_UPDATE_PASSWORD = 'profile_update_password',
  SOTORAGE_CHANGE = 'storage_change',
  SOTORAGE_READ_GLOBAL = 'storage_read_global',
  SOTORAGE_WRITE_GLOBAL = 'storage_write_global',
  SOTORAGE_READ_USER = 'storage_read_user',
  SOTORAGE_WRITE_USER = 'storage_write_user',
  USERS_ASSIGN_APP = 'users_assign_app',
  USERS_CREATE = 'users_create',
  USERS_DELETE = 'users_delete',
  USERS_INFO = 'users_info',
  USERS_LIST = 'users_list',
  USERS_UNASSIGN_APP = 'users_unassign_app',
  USERS_UPDATE = 'users_update',
  RB_CHANGE = 'rb_change',
  RB_GET = 'rb_get',
  RB_PUT = 'rb_put',
  RB_DELETE = 'rb_delete',
  RB_RESTORE = 'rb_restore',
  RB_CLEAN = 'rb_clean',
  APPS_LIST = 'apps_list',
  APPS_LIST_BY_UID = 'apps_list_by_uid',
  APPS_INSTALL = 'apps_install',
  APPS_UNINSTALL = 'apps_uninstall',
  APPS_CHANGE = 'apps_change',
  SOURCES_ENABLE = 'sources_enable',
  SOURCES_DISABLE = 'sources_disable',
  SOURCES_CHANGE = 'sources_change',
  SOURCES_GET = 'sources_get',
  PERMISSIONS_ENABLE = 'permissions_enable',
  PERMISSIONS_DISABLE = 'permissions_disable',
  PERMISSIONS_CHANGE = 'permissions_change',
  PERMISSIONS_GET = 'permissions_get'
}